[java-shell] Move console.log support to ConsoleLogSupport class
